prevent delete delegate if in meetings - committee status optimized - can add delegate validation rule changed

+++ Modules/Users/Http/Controllers/DelegateController.php
@@ -147,6 +147,10 @@ class DelegateController extends UserBaseController
 
     public function destroy(Delegate $delegate)
     {
+        if ($delegate->checkIfDelegateInMeetings())
+        {
+            return response()->json(['status' =>true, 'msg'=> __('users::delegates.delegate_in_meetings_can_not_remove')]);
+        }
         $delegate->log('delete_delegate');
         $delegate->delete();
         return response()->json(['msg' => __('users::delegates.deleted')]);
@@ -154,10 +158,16 @@ class DelegateController extends UserBaseController
 
     public function removeFromCommitte($delegate_id, $committee_id, $department_id, $reason)
     {
+        $delegate = Delegate::find($delegate_id);
         if (CommitteeDelegate::checkIfMainCoordinatorNominateDelegates($committee_id)) {
             return response()->json(['code' => '0', 'msg' => __('users::delegates.delegate_can_not_delete')]);
 
-        } else {
+        }
+        elseif ($delegate->checkIfDelegateInMeetings())
+        {
+            return response()->json(['code' => '0', 'msg' => __('users::delegates.delegate_in_meetings_can_not_remove')]);
+        }
+        else {
             $delegate = Delegate::find($delegate_id);
             $delegate->log('remove_delegate_from_committee');
             $delegate->removeDelegateFromCommittee($delegate, $committee_id, $department_id, $reason,true);
